refactor: refatora estrutura da classe Entidade para incluir Projetil como Entidade
